move util file from node into parent test folder
