+++ tests/node/test_control.py
@@ -1,8 +1,7 @@
 import pytest
-from tests.node.util import createNode, load_secrets_conf
+from tests.util import createNode, load_secrets_conf
 
 # Import Exceptions
-from defichain.exceptions.InternalServerError import InternalServerError
 
 node = createNode()
 address = load_secrets_conf()["wallet_address"]
