+++ tests/node/test_network.py
@@ -1,9 +1,8 @@
 import pytest
-from tests.node.util import createNode
+from tests.util import createNode
 
 # Import Exceptions
 from defichain.exceptions.InternalServerError import InternalServerError
-from defichain.exceptions.BadRequest import BadRequest
 
 node = createNode()
