+++ tests/node/test_zmq.py
@@ -1,5 +1,5 @@
 import pytest
-from tests.node.util import createNode
+from tests.util import createNode
 
 node = createNode()
